Rename spawn type ground to teleport
