+++ src/main/java/xiao/battleroyale/api/game/spawn/type/SpawnTypeTag.java
@@ -4,7 +4,7 @@ public class SpawnTypeTag {
 
     public static final String TYPE_NAME = "spawnType";
 
-    public static final String SPAWN_TYPE_GROUND = "ground";
+    public static final String SPAWN_TYPE_TELEPORT = "teleport";
     public static final String SPAWN_TYPE_PLANE = "plane";
 
     public static final String SPAWN_DETAIL = "detail";
